move ProgressBar to a standalone file (#4254)
