+++ cpp/open3d/Open3D.h.in
@@ -110,6 +110,8 @@
 #include "open3d/utility/ISAInfo.h"
 #include "open3d/utility/Logging.h"
 #include "open3d/utility/Parallel.h"
+#include "open3d/utility/ProgressBar.h"
+#include "open3d/utility/ProgressReporters.h"
 #include "open3d/utility/Timer.h"
 #include "open3d/visualization/gui/Application.h"
 #include "open3d/visualization/gui/Button.h"
