+++ cpp/open3d/io/TriangleMeshIO.cpp
@@ -30,6 +30,7 @@
 
 #include "open3d/utility/FileSystem.h"
 #include "open3d/utility/Logging.h"
+#include "open3d/utility/ProgressBar.h"
 
 namespace open3d {
 
@@ -107,7 +108,7 @@ bool ReadTriangleMesh(const std::string &filename,
         auto progress_text = std::string("Reading ") +
                              utility::ToUpper(filename_ext) +
                              " file: " + filename;
-        auto pbar = utility::ConsoleProgressBar(100, progress_text, true);
+        auto pbar = utility::ProgressBar(100, progress_text, true);
         params.update_progress = [pbar](double percent) mutable -> bool {
             pbar.SetCurrentCount(size_t(percent));
             return true;
