+++ cpp/open3d/io/file_format/FileSTL.cpp
@@ -31,6 +31,7 @@
 #include "open3d/io/TriangleMeshIO.h"
 #include "open3d/utility/FileSystem.h"
 #include "open3d/utility/Logging.h"
+#include "open3d/utility/ProgressBar.h"
 
 namespace open3d {
 namespace io {
@@ -77,8 +78,8 @@ bool WriteTriangleMeshToSTL(const std::string &filename,
     myFile.write(header, 80);
     myFile.write((char *)(&num_of_triangles), 4);
 
-    utility::ConsoleProgressBar progress_bar(num_of_triangles,
-                                             "Writing STL: ", print_progress);
+    utility::ProgressBar progress_bar(num_of_triangles,
+                                      "Writing STL: ", print_progress);
     for (size_t i = 0; i < num_of_triangles; i++) {
         Eigen::Vector3f float_vector3f =
                 mesh.triangle_normals_[i].cast<float>();
