+++ cpp/open3d/utility/ProgressReporters.h
@@ -26,7 +26,10 @@
 
 #pragma once
 
-#include "open3d/utility/Logging.h"  //for ConsoleProgressBar
+#include <functional>
+#include <string>
+
+#include "open3d/utility/ProgressBar.h"
 
 namespace open3d {
 namespace utility {
@@ -55,7 +58,7 @@ public:
         return CallUpdate(percent);
     }
     void Finish() { CallUpdate(100); }
-    // for compatibility with ConsoleProgressBar
+    // for compatibility with ProgressBar
     void operator++() { Update(last_count_ + 1); }
 
 private:
@@ -70,8 +73,7 @@ private:
     int64_t last_count_ = -1;
 };
 
-using utility::ConsoleProgressBar;
-/// update_progress(double percent) functor for ConsoleProgressBar
+/// update_progress(double percent) functor for ProgressBar
 struct ConsoleProgressUpdater {
     ConsoleProgressUpdater(const std::string &progress_info,
                            bool active = false)
@@ -85,7 +87,7 @@ struct ConsoleProgressUpdater {
     }
 
 private:
-    ConsoleProgressBar progress_bar_;
+    utility::ProgressBar progress_bar_;
     int last_pct_ = 0;
 };
