+++ examples/cpp/TrimMeshBasedOnPointCloud.cpp
@@ -82,7 +82,7 @@ int main(int argc, char* argv[]) {
     geometry::KDTreeFlann kdtree;
     kdtree.SetGeometry(*pcd);
     std::vector<bool> remove_vertex_mask(mesh->vertices_.size(), false);
-    utility::ConsoleProgressBar progress_bar(mesh->vertices_.size(),
+    utility::ProgressBar progress_bar(mesh->vertices_.size(),
                                              "Prune vetices: ");
 #pragma omp parallel for schedule(static) num_threads(utility::EstimateMaxThreads())
     for (int i = 0; i < (int)mesh->vertices_.size(); i++) {
